added solution hw-29

+++ src/main/java/com/internet/shop/controller/admin/DeleteUserController.java
@@ -1,7 +1,6 @@
 package com.internet.shop.controller.admin;
 
 import com.internet.shop.lib.Injector;
-import com.internet.shop.model.ShoppingCart;
 import com.internet.shop.service.ShoppingCartService;
 import com.internet.shop.service.UserService;
 import java.io.IOException;
@@ -11,6 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class DeleteUserController extends HttpServlet {
+    private static final String USER_ID = "user_Id";
     private static final Injector injector = Injector.getInstance("com.internet.shop");
     private UserService userService = (UserService) injector.getInstance(UserService.class);
     private ShoppingCartService cartService = (ShoppingCartService) injector
@@ -20,9 +20,15 @@ public class DeleteUserController extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         long userId = Long.parseLong(req.getParameter("id"));
+        long sessionUserId = (Long) req.getSession().getAttribute(USER_ID);
+        if (userId == sessionUserId) {
+            req.setAttribute("message", "You can't delete yourself!!!");
+            req.getRequestDispatcher(req.getContextPath() + "/users").forward(req, resp);
+            return;
+        }
         userService.deleteById(userId);
-        ShoppingCart userCart = cartService.getByUserId(userId);
-        cartService.delete(userCart);
+        cartService.getByUserId(userId)
+                .ifPresent(shoppingCart -> cartService.delete(shoppingCart));
         resp.sendRedirect(req.getContextPath() + "/users");
     }
 }
